feat: Add delete_folder_by_prefix function
This commit adds the `delete_folder_by_prefix` function to the `utils.py` file. This function allows for deleting a folder and its contents in the cloud storage service. It utilizes the `delete_resources_by_prefix` and `delete_folder` functions from the `cloudinary.api` module. This change enhances the functionality of the application by providing a convenient way to delete folders and their contents.