feat: Add API endpoint for creating user addresses
This commit adds a new API endpoint in the `address.py` module for creating user addresses. The `create_address` function handles the creation of a new address for the user, including validation of the address form data and setting the address as the default if specified. This enhancement improves the address management capabilities of the application.